bugfix: ockokit 라이브러리 이용을  api 라우터에서 서버 컴포넌트로 이동
